编写与运行集成测试
