--- test/test_driver/app_test.dart
@@ -1,19 +1,0 @@
-import 'package:flutter_driver/flutter_driver.dart';
-import 'package:test/test.dart';
-
-void main() {
-  group('App', () {
-    FlutterDriver driver;
-
-// 运行之前做的事情
-    setUpAll(() async {
-      driver = await FlutterDriver.connect();
-    });
-
-    tearDownAll(() async {
-      if (driver != null) {
-        driver.close();
-      }
-    });
-  });
-}
